pyproject.toml
[project]
name = "ragclassifier"
version = "0.1.0"
description = "LangGraph multi-agent document classifier with Streamlit UI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.50",
    "langchain-openai>=0.2.9",
    "langchain-core>=0.3.19",
    "streamlit>=1.39.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
]

